+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - [#9072](https://github.com/blockscout/blockscout/pull/9072) - Add tracing by block logic for geth
+- [#9056](https://github.com/blockscout/blockscout/pull/9056) - Noves.fi API proxy
 
 ### Fixes